+++ src/asm_unmarshal.cpp
@@ -295,7 +295,7 @@ struct Unmarshaller {
         res.pkt_access = proto.pkt_access;
         res.returns_map = proto.ret_type == Ret::PTR_TO_MAP_VALUE_OR_NULL;
         std::array<Arg, 7> args = {{Arg::DONTCARE, proto.arg1_type, proto.arg2_type, proto.arg3_type, proto.arg4_type, proto.arg5_type, Arg::DONTCARE}};
-        for (int i = 1; i < args.size() - 1; i++) {
+        for (size_t i = 1; i < args.size() - 1; i++) {
             switch (args[i]) {
             case Arg::DONTCARE:
                 return res;
